refactor(core): unify ui-library-context, delete 7 duplicates
Hoist UILibraryProvider/useUILibrary into @tour-kit/core, the new single
source of truth for the Radix vs Base UI library toggle. The seven copies
across react/hints/adoption/announcements/checklists/media/surveys diverged
only in JSDoc comments — the JSDoc'd react version is canonical.

- Move packages/react/src/lib/ui-library-context.tsx → packages/core
- Move regression test alongside the new canonical source
- Re-export UILibraryProvider, useUILibrary, UILibrary, UILibraryProviderProps
  from @tour-kit/core
- Delete 6 duplicate source files (-120 LoC of duplicated code)
- Repoint 13 consumer import sites to '@tour-kit/core'
- Drop dangling './lib/ui-library-context' re-exports from announcements'
  src/index.ts and src/headless.ts
- Add core barrel test + announcements barrel-cleanup test

All 1429 consumer tests pass; core test suite picks up the moved 7-case
regression test plus 3 new barrel assertions. Bundle delta: core +204B
(new file content), every consumer shrinks; net negative across workspace.
