+++ apps/docs/public/llms.txt
@@ -1,4 +1,4 @@
-# userTourKit v0.5.1 — Generated 2026-04-30T04:13:43Z
+# userTourKit v0.5.1 — Generated 2026-04-30T07:43:48Z
 
 # userTourKit
