+++ packages/adoption/src/components/feature-button.tsx
@@ -4,7 +4,7 @@ import { cn } from '@tour-kit/core'
 import * as React from 'react'
 import { useFeature } from '../hooks'
 import { Slot, UnifiedSlot } from '../lib/slot'
-import { useUILibrary } from '../lib/ui-library-context'
+import { useUILibrary } from '@tour-kit/core'
 import { type FeatureButtonVariants, featureButtonVariants } from './ui/button-variants'
 
 export interface FeatureButtonProps
