+++ packages/adoption/src/index.ts
@@ -47,7 +47,7 @@ export {
   useUILibrary,
   type UILibrary,
   type UILibraryProviderProps,
-} from './lib/ui-library-context'
+} from '@tour-kit/core'
 
 // ============================================
 // ENGINE (for advanced usage)
