+++ packages/announcements/src/components/announcement-close.tsx
@@ -4,7 +4,7 @@ import { cn } from '@tour-kit/core'
 import type { RenderProp } from '@tour-kit/core'
 import * as React from 'react'
 import { Slot, UnifiedSlot } from '../lib/slot'
-import { useUILibrary } from '../lib/ui-library-context'
+import { useUILibrary } from '@tour-kit/core'
 
 const defaultIcon = (
   <svg
