+++ packages/announcements/src/index.ts
@@ -128,7 +128,7 @@ export {
   useUILibrary,
   type UILibrary,
   type UILibraryProviderProps,
-} from './lib/ui-library-context'
+} from '@tour-kit/core'
 
 // ============================================
 // TYPES
