+++ packages/checklists/src/components/checklist-panel.tsx
@@ -4,7 +4,7 @@ import { cn } from '@tour-kit/core'
 import * as React from 'react'
 import { useChecklist } from '../hooks/use-checklist'
 import { Slot, UnifiedSlot } from '../lib/slot'
-import { useUILibrary } from '../lib/ui-library-context'
+import { useUILibrary } from '@tour-kit/core'
 import { Checklist } from './checklist'
 import { ChecklistProgress } from './checklist-progress'
 import {
