+++ packages/checklists/src/components/checklist.tsx
@@ -3,7 +3,7 @@
 import { cn } from '@tour-kit/core'
 import * as React from 'react'
 import { useChecklist } from '../hooks/use-checklist'
-import { useUILibrary } from '../lib/ui-library-context'
+import { useUILibrary } from '@tour-kit/core'
 import { ChecklistProgress } from './checklist-progress'
 import { ChecklistTask } from './checklist-task'
 import {
