+++ packages/hints/src/components/hint-hotspot.tsx
@@ -3,7 +3,7 @@
 import { cn } from '@tour-kit/core'
 import * as React from 'react'
 import { Slot, UnifiedSlot } from '../lib/slot'
-import { useUILibrary } from '../lib/ui-library-context'
+import { useUILibrary } from '@tour-kit/core'
 import type { HotspotPosition } from '../types'
 import { type HintHotspotVariants, hintHotspotVariants } from './ui/hint-variants'
