+++ packages/hints/src/components/hint-tooltip.tsx
@@ -15,7 +15,7 @@ import {
 import { cn } from '@tour-kit/core'
 import * as React from 'react'
 import { Slot, UnifiedSlot } from '../lib/slot'
-import { useUILibrary } from '../lib/ui-library-context'
+import { useUILibrary } from '@tour-kit/core'
 import type { Placement } from '../types'
 import {
   type HintTooltipVariants,
