+++ packages/media/src/index.ts
@@ -69,7 +69,7 @@ export {
   useUILibrary,
   type UILibrary,
   type UILibraryProviderProps,
-} from './lib/ui-library-context'
+} from '@tour-kit/core'
 
 export {
   parseMediaUrl,
