+++ packages/react/src/components/navigation/tour-close.tsx
@@ -4,7 +4,7 @@ import { useTour } from '@tour-kit/core'
 import { cn } from '@tour-kit/core'
 import * as React from 'react'
 import { type RenderProp, Slot, UnifiedSlot } from '../../lib/slot'
-import { useUILibrary } from '../../lib/ui-library-context'
+import { useUILibrary } from '@tour-kit/core'
 import { tourButtonVariants } from '../ui/button-variants'
 
 export interface TourCloseProps extends Omit<React.ComponentPropsWithoutRef<'button'>, 'children'> {
