+++ packages/react/src/index.ts
@@ -72,7 +72,7 @@ export {
   useUILibrary,
   type UILibrary,
   type UILibraryProviderProps,
-} from './lib/ui-library-context'
+} from '@tour-kit/core'
 
 // ============================================
 // HOOKS
